Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,10 +2,10 @@
 
 int main() {
     int a,b,c;
-    float avg;
+    int avg;
     scanf("%d %d %d",&a,&b,&c);
     avg=(a+b+c)/3;
-    printf("%f",avg);
+    printf("%.2f",avg);
 
     return 0;
 }
